+++ library/MultiRegions/GlobalLinSysPETScFull.h
@@ -69,7 +69,7 @@ public:
         const std::weak_ptr<ExpList> &pExpList,
         const std::shared_ptr<AssemblyMap> &pLocToGloMap);
 
-    MULTI_REGIONS_EXPORT ~GlobalLinSysPETScFull() override;
+    MULTI_REGIONS_EXPORT ~GlobalLinSysPETScFull() override = default;
 
 protected:
     /// Solve the linear system for given input and output vectors